+++ docs/CHANGELOG.md
@@ -2,6 +2,23 @@
  
 # Versions
  
+## 5.2.0
+- added support-utils for admin-server
+
+### new features
+- commons: added database-service with to execute common sql
+- commons: added admin-responses for admin-server+frontend
+
+### improvements
+- none
+
+### bug fixes
+- none
+ 
+### breaking changes
+- none
+
+
 ## 5.1.0
 - improved search-services